refactor: `schema` 리팩토링
