--- components/hospital/icu/main/chart/selected-chart/chart-body/chart-infos/cpcr-et-tube/cpcr-et-tube-schema.ts
@@ -1,6 +1,0 @@
-import { z } from 'zod'
-
-export const cpcrEtTubeSchema = z.object({
-  cpcr: z.string({ required_error: 'CPCR 여부를 선택해주세요' }),
-  etTube: z.string().nullable().optional(),
-})
