--- components/hospital/icu/main/chart/selected-chart/chart-body/chart-infos/group/group-schema.ts
@@ -1,7 +1,0 @@
-import { z } from 'zod'
-
-export const groupCheckFormSchema = z.object({
-  groupList: z.array(z.string()).refine((value) => value.some((item) => item), {
-    message: '적어도 하나의 그룹을 선택해주세요',
-  }),
-})
