+++ components/hospital/icu/main/chart/selected-chart/chart-body/table/order-form/drug-order/drug-form-field.tsx
@@ -3,7 +3,7 @@ import DrugComboBox from '@/components/hospital/icu/main/chart/selected-chart/ch
 import DrugDoseInput from '@/components/hospital/icu/main/chart/selected-chart/chart-body/table/order-form/drug-order/drug-dose-input'
 import DrugDoseUnitRadio from '@/components/hospital/icu/main/chart/selected-chart/chart-body/table/order-form/drug-order/drug-dose-unit-radio'
 import DrugSelectField from '@/components/hospital/icu/main/chart/selected-chart/chart-body/table/order-form/drug-order/drug-select-field'
-import { orderSchema } from '@/components/hospital/icu/main/chart/selected-chart/chart-body/table/order/order-schema'
+import { orderSchema } from '@/lib/schemas/icu/chart/order-schema'
 import { calculateTotalDrugAmount } from '@/lib/utils/utils'
 import { useCallback, useEffect, useState } from 'react'
 import { UseFormReturn } from 'react-hook-form'
