+++ lib/schemas/icu/chart/chart-info-schema.ts
@@ -1,4 +1,15 @@
-import * as z from 'zod'
+import { z } from 'zod'
+
+export const cpcrEtTubeSchema = z.object({
+  cpcr: z.string({ required_error: 'CPCR 여부를 선택해주세요' }),
+  etTube: z.string().nullable().optional(),
+})
+
+export const groupCheckFormSchema = z.object({
+  groupList: z.array(z.string()).refine((value) => value.some((item) => item), {
+    message: '적어도 하나의 그룹을 선택해주세요',
+  }),
+})
 
 export const vetsFormSchema = z.object({
   main_vet: z.string({ required_error: '주치의를 선택해주세요' }),
