+++ lib/schemas/icu/chart/order-schema.ts
@@ -14,3 +14,7 @@ export const orderSchema = z.object({
   icu_chart_order_comment: z.string().optional(),
   is_bordered: z.boolean().default(false).optional(),
 })
+
+export const ordererSchema = z.object({
+  orderer: z.string({ required_error: '오더를 내리는 수의사를 선택해주세요' }),
+})
