+++ lib/schemas/icu/chart/template-schema.ts
@@ -1,0 +1,16 @@
+import * as z from 'zod'
+
+export const templateFormSchema = z.object({
+  template_name: z
+    .string({ required_error: '템플릿 이름을 입력해주세요' })
+    .trim()
+    .min(1, {
+      message: '템플릿 이름을 입력해주세요',
+    }),
+
+  template_comment: z
+    .string({ message: '텍스트로 입력해주세요' })
+    .trim()
+    .optional()
+    .nullable(),
+})
